# backend/openapi_temp.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ZiggyAI Trading Platform",
    "description": "Complete AI-powered trading platform with market analysis, news, signals, and execution",
    "version": "1.0.0"
  },
  "paths": {
...
    "/screener/screener/scan": {
      "post": {
        "tags": ["screening", "screening"],
        "summary": "Screen Market",
        "description": "Screen market using cognitive core for high-quality signals.\n\nProcesses multiple symbols and returns those meeting criteria,\nsorted by signal quality and confidence.",
        "operationId": "screen_market_screener_screener_scan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ScreenerRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/app__api__routes_screener__ScreenerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
    "/screener/screener/presets/momentum": {
      "get": {
        "tags": ["screening", "screening"],
        "summary": "Momentum Screen",
        "description": "Pre-configured momentum screening.",
        "operationId": "momentum_screen_screener_screener_presets_momentum_get",
        "parameters": [
          {
            "name": "universe",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Universe to screen",
              "default": "sp500",
              "title": "Universe"
            },
            "description": "Universe to screen"
          },
          {
            "name": "min_confidence",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "description": "Minimum confidence",
              "default": 0.7,
              "title": "Min Confidence"
            },
            "description": "Minimum confidence"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
    "/screener/screener/regime_summary": {
      "get": {
        "tags": ["screening", "screening"],
        "summary": "Get Regime Summary",
        "description": "Get regime breakdown for a universe.",
        "operationId": "get_regime_summary_screener_screener_regime_summary_get",
        "parameters": [
          {
            "name": "universe",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Universe to analyze",
              "default": "sp500",
              "title": "Universe"
            },
            "description": "Universe to analyze"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ScreenerRequest": {
        "properties": {
          "universe": {
            "items": { "type": "string" },
            "type": "array",
            "title": "Universe",
            "description": "List of symbols to screen"
          },
          "min_confidence": {
            "type": "number",
            "title": "Min Confidence",
            "description": "Minimum signal confidence",
            "default": 0.6
          },
          "min_probability": {
            "anyOf": [{ "type": "number" }, { "type": "null" }],
            "title": "Min Probability",
            "description": "Minimum probability threshold"
          },
          "max_probability": {
            "anyOf": [{ "type": "number" }, { "type": "null" }],
            "title": "Max Probability",
            "description": "Maximum probability threshold"
          },
          "regimes": {
            "anyOf": [
              { "items": { "type": "string" }, "type": "array" },
              { "type": "null" }
            ],
            "title": "Regimes",
            "description": "Filter by specific regimes"
          },
          "sort_by": {
            "type": "string",
            "title": "Sort By",
            "description": "Sort by: confidence, probability, regime",
            "default": "confidence"
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "description": "Maximum results to return",
            "default": 50
          }
        },
        "type": "object",
        "required": ["universe"],
        "title": "ScreenerRequest",
        "description": "Request for market screening."
      },
      "ScreenerResult": {
        "properties": {
          "symbol": { "type": "string", "title": "Symbol" },
          "p_up": { "type": "number", "title": "P Up" },
          "confidence": { "type": "number", "title": "Confidence" },
          "regime": { "type": "string", "title": "Regime" },
          "top_features": {
            "items": { "items": {}, "type": "array" },
            "type": "array",
            "title": "Top Features"
          },
          "score": { "type": "number", "title": "Score" },
          "position_size": {
            "anyOf": [
              { "additionalProperties": true, "type": "object" },
              { "type": "null" }
            ],
            "title": "Position Size"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "p_up",
          "confidence",
          "regime",
          "top_features",
          "score"
        ],
        "title": "ScreenerResult",
        "description": "Screener result for a single symbol."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
      "app__api__routes_screener__ScreenerResponse": {
        "properties": {
          "results": {
            "items": { "$ref": "#/components/schemas/ScreenerResult" },
            "type": "array",
            "title": "Results"
          },
          "total_screened": { "type": "integer", "title": "Total Screened" },
          "filters_applied": {
            "additionalProperties": true,
            "type": "object",
            "title": "Filters Applied"
          },
          "execution_time_ms": {
            "type": "number",
            "title": "Execution Time Ms"
          },
          "regime_breakdown": {
            "additionalProperties": { "type": "integer" },
            "type": "object",
            "title": "Regime Breakdown"
          }
        },
        "type": "object",
        "required": [
          "results",
          "total_screened",
          "filters_applied",
          "execution_time_ms",
          "regime_breakdown"
        ],
        "title": "ScreenerResponse",
        "description": "Response from screener."
      },
...
    }
  }
}